sensor-relay-project/requirements.txt
pyserial
tkinter